requirements.txt
opencv-python
python-osc
mediapipe
pyserial
numpy